Updated hgather files
Fixed several issues with the addon.

- Changed the variables for the chat to monitor.
This fixed it to where certain items would never be counted because of the wording. The problem existed when it was matching the message wording.
- Moved the variables outside of the if statement to capture the text before the if statement. 
This fixed the issue to where the count would not be exact. The problem was if a different message came in right after the item was gotten, it would take the new message instead of the intended message.
- Added a percentage to the break counter for mining.  
This enhances the addon to show how often breaks will happen. 
- Added a percentage to the items gathered for mining.
This enhances the addon to show how likely you are to get an item.
- Updated the item list for the tracker.
Added more items to the list that were not there before. 
- Cleaned up the code a little due to the changes that were made. 
I removed several things from the addon that were redundant after the changes that were implemented.